projetos/Meu modelo flask deploy heroku/requirements.txt
flask
gunicorn
python_dotenv